@@ -41,7 +41,7 @@ ActiveRecord::Schema.define do
   end
 end
 
-# prepare rails app
+# prepare Rails app
 require 'action_controller/railtie'
 require 'action_view/railtie'
 require 'active_admin'
