@@ -1,8 +1,6 @@
 ActiveAdmin.register Store do
 
-  if Rails::VERSION::MAJOR >= 4
-    permit_params :name
-  end
+  permit_params :name
 
   index pagination_total: false
 
